% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Badanie wyników gry}
\usage{
badaj_gre(strategia, powt = 100, progressBar = FALSE, histogram = FALSE,
  statystyki = TRUE, zwrocRozklad = FALSE)
}
\arguments{
\item{strategia}{Strategia, którą będziemy testować.}

\item{powt}{Liczba badanych gier.}

\item{progressBar}{Gdy TRUE wyświetlany jest pasek postępu, gdy FALSE paska nie ma.}

\item{histogram}{Gdy TRUE wyświetlany jest histogram wyników, gdy FALSE histogramu nie ma.}

\item{statystyki}{Gdy TRUE wyświetlane są statystyki rozkłądu wyników, gdy FALSE to nie.}

\item{zwrocRozklad}{Gdy TRUE funkcja zwraca nazwany wektor statystyk opisowych.}
}
\description{
Funkcja powtarzającą model gry SuperFarmer z zadaną strategią 10000 razy i wyznaczającą podstawowe statystyki opisowe rozkładu wyników
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

