% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czyWygrana.R
\name{czyWygrana}
\alias{czyWygrana}
\title{Weryfikacja warunków końca gry}
\usage{
czyWygrana(stanGracza)
}
\arguments{
\item{stanGracza}{Nazwany wektor liczebności zwierząt w zagrodzie gracza.
Nazwy zwierząt i kolejność w wektorze: c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")}
}
\value{
Funkcja zwraca wartość 1 gdy nastąpił koniec gry lub 0 gdy gra ma trwać kolejną rundę.
}
\description{
Niniejsza funkcja przeprowadza test, czy nastąpił koniec gry SuperFarmer,
tj. czy gracz uzyskał przynajmniej po jednym zwierzęciu
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

