% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodajZwierzetaDoZagrody.R
\name{dodajZwierzetaDoZagrody}
\alias{dodajZwierzetaDoZagrody}
\title{Transfer zwierząt do zagrody}
\usage{
dodajZwierzetaDoZagrody(stanGry, wynikiRzutuKoscmi)
}
\arguments{
\item{stanGry}{Lista dwóch nazwanych wektorów opisujących aktualny stan gry poprzez podanie liczebności zwierząt w stadzie gracza oraz liczebności zwierząt w banku. Wymagane nazwewnictwo i kolejność zwierząt w wektorach to c("krolik", "owca", "swinia", "krowa", "kon", "maly_pies", "duzy_pies").}

\item{wynikRzutuKoscmi}{Dwuelementowy wektor przyjmujący jako wartości nazwy zwierząt, które zostały wylosowane przy rzucie koścmi.}
}
\description{
Niniejsza funkcja przeprowadza operację wymiany zwierząt pomiędzy stadem głównym a stadem gracza
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

