% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.r
\name{gra}
\alias{gra}
\title{Gra SuperFarmer na 1 os. wyposażoną w strategię}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{Funkcja opisująca zbiór reguł, które reagują na aktualny stan stada gracza i zwraca nowy stan stada po wykonaniu wszystkich wymian zgodnie z zasadami gdy}
}
\value{
Funkcja zwraca liczbę tur, jaka była niezbędna do zakończenia gdy przy zastosowanej strategii
}
\description{
Niniejsza funkcja ma na celu przeprowadzenie pełnej gry SuperFarmer dla jednego gracza, który posługuje się strategią zdefiniowaną jako argument funkcji.
}
\examples{
gra(strategia_Pierwsza)

}

