% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozegrajGre.R
\name{rozegrajGre}
\alias{rozegrajGre}
\title{Pojedyncza rozgrywka w SuperFarmera}
\usage{
rozegrajGre(strategia, stanGry)
}
\arguments{
\item{strategia}{Funkcja opisująca zbiór reguł, które reagują na aktualny stan stada gracza i zwraca nowy stan stada po wykonaniu wszystkich wymian zgodnie z zasadami gry.}

\item{stanGry}{Lista dwóch nazwanych wektorów opisujących aktualny stan gry poprzez podanie liczebności zwierząt w stadzie gracza oraz liczebności zwierząt w banku. Wymagane nazwewnictwo i kolejność zwierząt w wektorach to c("krolik", "owca", "swinia", "krowa", "kon", "maly_pies", "duzy_pies").}
}
\value{
Liczba tur, jakie były niezbędne do zakończenia gry od zadanego stanu gry, przy wybranej strategii.
}
\description{
Funkcja przeprowadzająca jedną, jednoosobową grę SuperFarmer do momentu wygranej gracza przy zadanej strategii oraz zadanego stanu początkowego gry.
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

