% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzucKoscmi.R
\name{rzucKoscmi}
\alias{rzucKoscmi}
\title{Implementacja rzutu kośćmi}
\usage{
rzucKoscmi()
}
\arguments{
\item{None}{}
}
\value{
Funkcja zwraca dwuelementowy wektor wyników, który jako wartości przyjmuje nazwy zwierząt zgodnie
z przyjętą konwencją nazewnictwa: c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies","wilk","lis")
}
\description{
Niniejsza funkcja przeprowadza losowanie dwoma kostkami do gry SuperFarmer. Specyfikacja kostek zgodna
z instrukcją gry:
Kostka zielona: 6 x Królik, 3 x Owca, 1 x Świnia, 1 x Krowa, 1 x Mały pies
Kostka czerwona: 6 x Królik, 2 x Owca, 1 x Krowa, 1 x Duży pies
}
\examples{
rzucKoscmi()

}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

