% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_Druga.R
\name{strategia_Druga}
\alias{strategia_Druga}
\title{Strategia Druga}
\usage{
strategia_Druga(stanGracza)
}
\arguments{
\item{stanGracza}{Nazwany wektor liczebności zwierząt w zagrodzie gracza.
Nazwy zwierząt i kolejność w wektorze: c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")}
}
\value{
Strategia zwraca wektor stanGracza po wykonaniu zmian zapisanych w strategii
}
\description{
Druga zaimplementowana strategia gry jednego gracza w SuperFarmera.
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

