% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyliczStanGryPoRzucieKoscmi.R
\name{wyliczStanGryPoRzucieKoscmi}
\alias{wyliczStanGryPoRzucieKoscmi}
\title{Wyliczenie stanu gry po rzucie kośćmi}
\usage{
wyliczStanGryPoRzucieKoscmi(stanGry)
}
\arguments{
\item{stanGry}{Lista dwóch nazwanych wektorów opisujących aktualny stan gry poprzez podanie liczebności zwierząt w stadzie gracza oraz liczebności zwierząt w banku. Wymagane nazwewnictwo i kolejność zwierząt w wektorach to c("krolik", "owca", "swinia", "krowa", "kon", "maly_pies", "duzy_pies").}
}
\value{
Funkcja zwraca zaktualizowaną listę stanGry opisującą stan stada gracza oraz stan stada głównego
}
\description{
Niniejsza funkcja aktualizuje stan stada gracza oraz stan głównego stada po wykonaniu losowania kostkami.
}
\author{
Tomasz Bielak \email{tomasz.bielak@student.uw.edu.pl}

Michał Kotlewski \email{tomasz.bielak@student.uw.edu.pl}
}

